backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0
google-generativeai==0.3.2
chromadb==0.5.23
numpy<2.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
apscheduler==3.10.4
aiosqlite==0.19.0
pydantic>=2.5.3
httpx>=0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
